requirements.txt
APScheduler==2.1.2
Flask==0.10.1
Flask-Bcrypt==0.5.2
Flask-Login==0.2.7
Flask-OAuth==0.12
Flask-SQLAlchemy==1.0
Jinja2==2.7.1
MarkupSafe==0.18
MySQL-python==1.2.4
SQLAlchemy==0.8.3
Werkzeug==0.9.4
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
oauth==1.0.1
oauth2==1.5.211
psycopg2==2.5.1
py-bcrypt==0.4
python-crontab==1.6.0
python-dateutil==2.2
pytz==2013.9
redis==2.9.1
rq==0.3.13
simplejson==3.3.1
six==1.4.1
times==0.6.2
wsgiref==0.1.2
mailchimp=2.0.9
Flask-API=0.6.1
apscheduler=3.0.1
